v1.0.3 发送网络实时状态,紫外线指数改为实时级别，停止传感器数据的实时传输

+++ app/src/main/java/com/example/messengerservicedemo/response/DailyResponse.kt
@@ -21,6 +21,6 @@ class DailyResponse(val status: String, val result: Result) {
 
     class LifeIndex(val coldRisk: List<LifeDescription>, val carWashing: List<LifeDescription>, val ultraviolet: List<LifeDescription>, val dressing: List<LifeDescription>)
 
-    class LifeDescription(val desc: String,val index: Float)
+    class LifeDescription(val desc: String,val index: Int)
 
 }
